Add mdcodes npm package and codelist service
Add ember-browserify to support npm module imports and add mdcodes dependency.

Take the codelists from the mdcodes npm module and make them available
as a service. This allows the potential to swap out or modify the
codelists by overriding the object returned by the service. The
individual codelist names are re-mapped to be more generic, e.g. "role"
instead of "iso_role". For convenience during development, *all* of the lists
are being imported, the build size could be reduced by only importing the
ones actually utilized in the editor.

Closes: #12