Implement copy and delete actions for contacts
Adds a new contact/new/id route that will generate a new contact if not
supplied a contact record via a transition. The copied contact will
first be displayed in the "new" route, prompting the user to change the
name(s).

The delete action on the contact/show/edit route is now functional.