Handle unimplemented sections
This commit makes the mdJSON service aware of uminplemented properties
of the mdJSON schema. This is handled by defining an array of mdJSON
object paths that correspond to unimplemented properties. These
properties are then set to *undefined* before generating the mdJSON
(these *undefined* properties are actually removed when the JSON
objected is cleaned).

Fixes #119