Refactor new item routes
Potential data loss can occur when creating new records. The bug can be
triggered by entering the new/id route and migrating away from the route
without saving the record. The record is not properly removed from the
store, resulting in an inconsistent state where the record appears to
still exist(if the route is accessed using the back button).
Furthermore, the new/index routes forward to the new/id routes after
instantiating, but not saving, a new record. The transition is provided
the newly created model, therefore the model hook is never fired on the
new/id route. Supplying the model id instead will ensure that the model
hook is fired.

Additionally, the base model should not attempt autoSave the record if
it is in an `empty` state.

The record model is setting the allRecords property in the init method.
This appears  to be causing the record to remain locked in the `new`
state and binding is broken. Assigning allRecords to a variable in the
computed properties appears to resolve this(at least the bug is not
reproducible).

Closes #306