Feature/auto load vocabularies by profile (#564)
* Add vocabulariesUrl Environment Variable

* Install Axios

Added axios as a dependency.
Also updating yarn.lock.

* Add Dynamic Loading Thesauri

Deprecated/removed the gcmd-keywords and mdKeywords modules.
Loading the list of vocabularies dynamically from the config file in
mdKeywords repository.
Using the urls from that config file, loading the keywords dynamically
from other config files in the mdKeywords repository.

* Add Tooltip to Thesauri Select Input Component

Note that the tooltipPath value in the template file needs to be the
same as the name of the tooltipText attribute in the component file.
The text inside the tooltip comes from the "description" attribute in
the citation object for the thesaurus.

* Update Yarn Lock

Removing gcmd-keywords and mdKeywords from yarn.lock.

* Update Environment Config

Changed the name of the vocabularies file in mdKeywords. Needed to also
be updated in this environment config.

* Add Feature

Record route retrieves list of profiles using a file in mdProfiles
repository and loads it into the store.
Profile service adapts the profiles from the store to be used in the
service.
Profile model updated.
Deprecated / removed mdProfiles module from package.json.
Updated yarn.lock.

* Update Package

* Fix Import

Add axios import and profilesListUrl.

* Move Profiles List Url to Environment Config

* Fix Reference to Profiles List Url

* Add Feature

Vocabularies are auto-loaded to the record based on the selected profile

* Remove Console Log

* Clean Up Env Config

* Fix Adapter

* Fix Model

* New Feature: Profile Switching

Added the ability to add/remove vocabularies to/from a record when
switching the profile.

* Clean Up

* Update Env Vars

Updating the profiles and vocabularies lists environment
variables to point to updated files.

* Update Env Var

Using a different profiles list file.

* Refactor

Replacing the adpater function with toJSON,
the proper way to convert a class to an object.

* Revert Changes

Removing the auto-add logic

* Reverting Changes

Removed ;

* Reverting Changes

* Revert Changes

* Refactor Load Profiles

* Cleanup

* Replacing Custom Profile Logic

* Update Env Vars

* Fix Env Var

* Filter Thesauri

New custom profiles include a vocabularies section,
the list of thesauri is filtered based on that list.
If the profile does not include a vocabularies section
it will filter based on the default vocabularies.

* Cleanup

* Cleanup Imports

* Update Config

Change master to dev

* Fix Typo

* Bugfix Profile and Vocabularies

Neither profiles nor vocabularies were being loaded early enough,
so each one had to be moved into the application route.
This is better because it ensures they load before any routes
are displayed.